# Conflicts:
#	Morty/Demo/Demo.cpp
#	Morty/Demo/Test/BasicShape.h
#	Morty/Demo/Test/BasicTransform.h
#	Morty/Demo/Test/BoneAnimation.h
#	Morty/Demo/Test/EnvironmentCubemap.h
#	Morty/Demo/Test/Floor.h
#	Morty/Demo/Test/GPUDrivenCulling.h
#	Morty/Demo/Test/LoadModel.h
#	Morty/Demo/Test/LoadModel_Sponza.h
#	Morty/Demo/Test/Pbr.h
#	Morty/Demo/Test/PointLight.h
#	Morty/Demo/Test/ShadowMap.h
#	Morty/Demo/Test/WaveToRGB.h
#	Morty/Editor/Property/PropertyBase.cpp
#	Morty/Render/Culling/MGPUCameraFrustumCulling.cpp
#	Morty/Render/Material/MMaterial.cpp
#	Morty/Render/Material/MMaterial.h
#	Morty/Render/Material/MShader.h
#	Morty/Render/Material/MShaderBuffer.cpp
#	Morty/Render/Material/MShaderParam.h
#	Morty/Render/Model/MAnimationRenderProxy.cpp
#	Morty/Render/Model/MModelConverter.cpp
#	Morty/Render/Render/MRenderCommand.h
#	Morty/Render/Render/MRenderGlobal.cpp
#	Morty/Render/Render/MRenderGlobal.h
#	Morty/Render/Render/MVertex.h
#	Morty/Render/Render/Vulkan/MVulkanPipelineManager.cpp
#	Morty/Render/Render/Vulkan/MVulkanRenderCommand.cpp
#	Morty/Render/Render/Vulkan/MVulkanRenderCommand.h
#	Morty/Render/Render/Vulkan/MVulkanShaderCompiler.cpp
#	Morty/Render/RenderProgram/MDeferredRenderProgram.cpp
#	Morty/Render/RenderProgram/MFrameShaderPropertyBlock.cpp
#	Morty/Render/RenderProgram/MeshRender/MCullingResultRenderable.cpp
#	Morty/Render/RenderProgram/RenderWork/MDeferredLightingRenderWork.cpp
#	Morty/Render/RenderProgram/RenderWork/MPostProcessRenderWork.cpp
#	Morty/Render/RenderProgram/RenderWork/MTransparentRenderWork.cpp
#	Morty/Render/RenderProgram/RenderWork/MVoxelizerRenderWork.cpp
#	Morty/Render/Shadow/MShadowMapShaderPropertyBlock.cpp
#	Morty/Render/Shadow/MShadowMapUtil.cpp
#	Morty/Render/Shadow/MShadowMeshManager.cpp
#	Resource/Shader/Deferred/deferred_gbuffer.mps
#	Resource/Shader/Deferred/deferred_lighting.mps
#	Resource/Shader/Forward/light_forward.hlsl
#	Resource/Shader/Forward/model.mvs
#	Resource/Shader/Internal/internal_constant.hlsl
#	Resource/Shader/Internal/internal_uniform_model.hlsl
#	Resource/Shader/Lighting/pbr_lighting.hlsl
#	Resource/Shader/Lighting/pbr_uniform_material.hlsl
#	Resource/Shader/Model/universal_model.mvs
#	Resource/Shader/PostProcess/post_process_basic.mvs
#	Resource/Shader/Shadow/shadowmap.mvs
#	Resource/Shader/Voxel/voxel_debug_view.mps
#	Resource/Shader/Voxel/voxel_debug_view.mvs
#	Resource/Shader/Voxel/voxelizer.mgs
#	Resource/Shader/Voxel/voxelizer.mps
#	Resource/Shader/Voxel/voxelizer.mvs

+++ Morty/Core/Utility/MMemoryPool.cpp
@@ -18,7 +18,7 @@ MMemoryPool::~MMemoryPool()
 
 }
 
-bool MMemoryPool::AllowMemory(const uint32_t& unVariantSize, MemoryInfo& info)
+bool MMemoryPool::AllowMemory(const size_t& unVariantSize, MemoryInfo& info)
 {
 	auto biggestIter = m_vFreeMemory.begin();
 	auto bestIter = m_vFreeMemory.end();
